sync single repo data
