+++ src/app/api/webhooks/github/route.ts
@@ -7,6 +7,7 @@ import { createClient } from "@/utils/supabase/server";
 import { Json, TablesInsert } from "@/types/supabase";
 
 import * as schemas from "./schemas";
+import { createNewRepo } from "./route/utils";
 
 export async function POST(request: Request) {
   const cookieStore = cookies();
@@ -64,23 +65,7 @@ export async function POST(request: Request) {
     let finalRepoData = repoData;
     if (!repoData) {
       // Create the repository and repositories_secure entries
-      const secret = process.env.GITHUB_WEBHOOK_SECRET;
-      const { data: newRepo } = await supabase
-        .from("repositories")
-        .insert({ org, repo })
-        .select("id")
-        .single()
-        .throwOnError();
-      await supabase
-        .from("repositories_secure")
-        .insert({ repo_id: newRepo.id, secret })
-        .throwOnError();
-      const { data: fetchedRepo } = await supabase
-        .from("repositories")
-        .select("id, repositories_secure ( secret )")
-        .eq("id", newRepo.id)
-        .single();
-      finalRepoData = fetchedRepo;
+      finalRepoData = await createNewRepo({ supabase, org, repo });
     }
 
     if (!finalRepoData?.repositories_secure?.secret)
